changes styles for register-form
